SQL-DB Connections
Connected mySQL database server to our application. Added a new project called LibraryData which holds all the packages to connect to the database. Inside the LibraryData project, there is a c# file name "LibraryContent" which creates an object database "Patron."  Inside Model directory, there is a file name "Patron.cs" which creates a Data Model with all the necessary columns for our application. Our Team450Project references the LibraryData project. Services is to add DBSet is done inside the startup file. Lastly, using migration we initialized the table in SQL.